requirements.txt
# This file lists the direct dependencies of the project with pinned versions
# to ensure reproducible builds for a production environment.

# --- Core Application Framework ---
fastapi==0.116.1
uvicorn[standard]==0.35.0

# --- AI & RAG Core Dependencies ---
# Using Google Gemini as the LLM
langchain==0.3.26
langchain-core==0.3.69
langchain-community==0.3.27
langchain-google-genai==2.1.8

# --- Local Embeddings & Reranking ---
# Running embeddings on the application server.
# This requires torch and transformers, which are large libraries.

FlashRank==0.2.10

# --- Vector Store Client ---
# Using Pinecone for the vector database.
langchain-pinecone==0.2.9

# --- Data Ingestion & Utilities ---
# For loading documents from S3-compatible storage (Cloudflare R2).
boto3==1.39.7
langchain-aws==0.2.28
# For loading and parsing specific file types.
pypdf==5.7.0
unstructured==0.18.5
# For loading environment variables from a .env file.
python-dotenv==1.1.1

# --- Authentication ---
# For handling JWTs and password hashing.
# Explicitly pinning all components for security and stability.
passlib==1.7.4
bcrypt==4.3.0
python-jose==3.5.0
cryptography==45.0.5

# --- Custom Features ---
# For the ticket suggestion feature.
numpy

# --- Monitoring & Observability (Optional) ---
# For tracing and debugging RAG pipelines with LangSmith.
# Can be removed for a minimal deployment.
langsmith==0.4.5